fix: have Screenshot call super method to get attr defined

+++ bot_utils/picture_in_picture/picture_input/picture_input_abc.py
@@ -12,4 +12,5 @@ class PictureInputAbstractClass(DebugAbstractClass):
         self.img_path = img_path
 
     def clean(self):
-        os.remove(self.img_path)
+        if self.img_path:
+            os.remove(self.img_path)
